requirements.txt
GitPython>=3.1.40
pydantic>=2.0.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
PyYAML>=6.0
click>=8.1.0
tabulate>=0.9.0
colorama>=0.4.6
